[fix] get negative data with big int

+++ libs/libvnstat/src/db/models/traffic.rs
@@ -1,4 +1,4 @@
-use diesel::sql_types::{Integer, Text};
+use diesel::sql_types::{BigInt, Integer, Text};
 use serde::Serialize;
 
 #[derive(Debug, QueryableByName, Serialize, Clone)]
@@ -9,8 +9,8 @@ pub struct Traffic {
     pub interface: i32,
     #[sql_type = "Text"]
     pub date: String,
-    #[sql_type = "Integer"]
-    pub rx: i32,
-    #[sql_type = "Integer"]
-    pub tx: i32,
+    #[sql_type = "BigInt"]
+    pub rx: i64,
+    #[sql_type = "BigInt"]
+    pub tx: i64,
 }
